+++ src/i_video.c
@@ -672,7 +672,7 @@ void I_RestoreMousePointerPosition(void)
 
 static void SmoothMouse(int *x, int *y)
 {
-    const fixed_t   tic = (((int64_t)I_GetTimeMS() * TICRATE) % 1000) * FRACUNIT / 1000;
+    const fixed_t   tic = ((I_GetTimeMS() * TICRATE) % 1000) * FRACUNIT / 1000;
     const fixed_t   adjustment = FixedDiv(tic, FRACUNIT + tic);
     static int      xx, yy;
